+++ CHANGES.md
@@ -14,6 +14,7 @@
 - FIX: Don't try to read KML with `pyogrio <= 0.11.0 `
 - FIX: Fix `ci.assert_val` with nans
 - FIX: Better nan management in `ci.assert_raster_almost_equal_magnitude`
+- FIX: Make vectorize-based functions to work with `xarray.Dataset`
 - CI: Correctly compute numpy case with dask_env decorator
 - CI: Correctly manage dask in `test_rasters.test_set_nodata` and `test_rasters.test_where` (create a dask array to evaluate laziness)